move filename creation code to its own file
